Move debugger setup code out of `lib.rs`
